correct the restore of vnode mappings and do some refactors using macro
